agent/pyproject.toml
[project]
name = "echopay-agent"
version = "0.1.0"
description = "EchoPay AI agent — FastAPI service powering the bunq payment loop"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.*",
  "uvicorn[standard]==0.32.*",
  "pydantic==2.9.*",
  "python-dotenv==1.0.*",
  "python-multipart==0.0.*",
  "httpx==0.27.*",
  "anthropic==0.39.*",
  "supabase==2.8.*",
  "voyageai==0.3.*",
  "pyjwt==2.9.*",
  "pytest==8.*",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
